debian: consolidate source and binary repositories
This way, source and binary packages are published at the same url.

Move the creation of the source package into release/release-debian to
have access to the same repoository.

Source packages are now also signed, which requires the changelog entry
to be signed with the cockpitous email address instead of the author of
the tag.

This change depends on the new debian package in cockpit, which creates
a non-native package (using the original tarball and a separate tarball
for the debian directory).