release-debian: fix skipping existing builds
Now that we include source packages in the repository, `reprepro
listfilter` returns multiple architectures (amd64 and source) for the
cockpit package. Thus, the condition was always true and we never
skipped the build.

Fix this and make it future proof by only building when listfilter
returns nothing for a given $TAG.