release: Add a release-source command
A script that takes a git repo in the current working directory and builds
a tarball from a tag plus full tarball patches for each commit up until
the current HEAD. The patch files include files generated by autotools
or dist/ directories.

The output patchset is in git patch format with the intent that it can
contain binary patches. The resulting patch file names are placed in the
current directory and their names are printed on stdout.