release: Add a release-srpm command
A script that creates an SRPM based on a tarball and patches.
The spec file is specified on the command line. It is updated with
version and patch information as well as changelog entries.

If current working directory is a git repository and a tag is specified.
then the changelog is updated based on the text in that git tag.