sink: Remove badge support from sink
This is a workaround for the fact that there was no service hosting
image badges for specific github statuses. We no longer use this.

If anyone ever does need this again, the correct way to solve this
is to host a service that looks up the GitHub status for master
and produces an image based on it.

Closes #60