tests: Consistent invocation and container mount points
The verify container is now called cockpit/tests to match
the cockpit-tests RPM and general expectation that this is
something more and more developers start to use.

The container can be invoked with the following command, and
it will be ready to run the integration tests.

 $ sudo atomic run cockpit/tests

It can be installed as a service on a system with the following
command:

 $ sudo atomic install cockpit/tests

In addition we define standard mount points

 * /secrets: A directory containing SSH or github-token secrets
 * /images: An images directory

Added a 'make tests-secrets' target which wraps up the necessary
secrets both in Kubernetes Secrets JSON.

Added Openshift deployment info and JSON files.

Closes #101