tests: Run or load the .tasks file in the target repository
This file produces the possible set of bot tasks to be executed in the
current repository by bots. Empty lines and commented lines are ignored
and commented lines are removed. If this file is executable it will be
run, and the output will be used as the list of tasks.

The tasks are sorted descending alphabetically and a task is chosen from
near the top of the sorted result. The task will be run in a shell by the
Cockpit bots.

Closes #120
Reviewed-by: Martin Pitt <martin@piware.de>