tests: Add cleanup hacks for container networking
The first hack is because we don't want to see this error message
when libvirt tries to write bridge related stuff:

error: Unable to set bridge cockpit1 forward_delay: Read-only file system

Secondly, when run in Kubernetes the Pod can outlast the Container and
so this network may have leaked in from elsewhere. Lets double
check and remove it using brctl.

Neither of these hacks work in the tasks code because it may be
executing older branches which do not account for these things.

Closes #131