release: Make release-source preseed rebuild with tarballs
The tarballs of a distributed package contain some random
decisions about dependency versions, configure and aclocal.m4
contants and other sources of variation.

In order for release-source to create small reliable patches
we should preseed our rebuild directory with the tarballs.

This only affects files that are either ignored by git or
are otherwise extra in the repo. Committed files remain
true to git commits and are not seeded from tarballs.