images: Expose http on standard port 80
Cockpit expects to talk to cockpit-11 on the standard HTTP port [1], and
apparently the created cockpit-tests.service had been manually changed
for that until the reinstall. As we don't need port 80 for anything
else, change the port in the Dockerfile instead of changing REDHAT_PING,
to keep the recent behaviour.

[1] https://github.com/cockpit-project/cockpit/blob/master/bots/tests-scan#L58