release: Drop release-debian
This script is cockpit specific. It currently pushes packages to a
custom repository which is not being advertised any more (since cockpit
is now in Debian proper), and also to jessie which is EOL soon and not
getting tested in CI any more.

Automating proper Debian downstream releases needs to be reworked and
rethought. Until then, drop this script to avoid wasting lots cycles on
releases.