release: Replace release polling with GitHub trigger
So far cockpit-release.service restarted itself every 30 minutes to
poll if a new release is pending. This is a waste most of the time, and
an annoying delay when a release actually happens.

Make this unit start via a HTTP request instead, so that it becomes
suitable as a GitHub "create" web hook, and generalize the mechanism to
work with more than just cockpit:

Add a releasetrigger.socket that listens on port 8090 and activates
releasetrigger@.service. This receives a HTTP request and uses the
path to select the project (sanitizing it to just alphanumeric
characters). If `<project>-release.service` exists, start that and
report success, otherwise report a 404.

There is no user input aside from the sanitized HTTP path, so at most a
remote can try to DoS. Mitigate this by limiting cockpit-release.service
to three starts a day, and if it is already running trying to start it
again is a no-op.

Drop the `make release-install` dependency to `release-container`. There
is no need to build a container image, we want to user the one from
dockerhub, as documented in README.

Update the documentation accordingly. Refer to `make release-install`
instead of replicating the setup commands.

Closes #175