release: Allow adding dynamic user to passwd
In OpenShift environments with a randomized UI (such as CentOS CI),
allow that user to add itself to `/etc/passwd`. The `ssh` client
requires a valid passwd entry.

This is the recommended solution in
https://docs.openshift.com/container-platform/3.7/creating_images/guidelines.html

This does not affect running the container directly (e. g. through the
systemd unit), as the `user` user is not in the `root` group.