release: Add resources and webhook for running in OpenShift
Add a script to turn a home directory with all the necessary credentials
into a Kubernetes secrets volume.

Add a GitHub webhook (using only Python 3 included modules) that can
release any project with the Cockpituous release script. The git repo
URL is taken from the requests's GitHub metadata, and the path to the
release script is taken from the URL path.

Document how to deploy this.

Also install the `oc` client into the container. It is not currently
being used, but it makes it easier to experiment with different
architectures, such as creating `Job` objects in a separate webhook
container.

Closes #183