release: Don't force ssh TTY
This causes eternal hangs, as an EOF on stdin is then not sufficient any
more to shut down ssh.  This can easily be demonstrated with

   echo whoami | ssh -tt somehost

This was introduced in commit 3f2c177abfff for release-debian. But this
has been gone for a  while, and there is no real justification to do
that -- nothing in our release scripts ought to need a PTY, and if there
is anything that should be wrapped in `script` or similar.

Closes #188