release: Drop support for systemd socket activation webhook
These got replaced with the Kubernetes route based approach. There was
no authentication, and we don't test this mode any more in production,
so drop it.

Provide a manual fallback through `make` targets `release-cockpit` and
`release-shell`, in case our CentOS CI release deployment is not
available, or for debugging some steps of the release.

Also fix heading levels in the documentation.