release: Support spec files with pre-filled Version:
starter-kit based projects do release tarballs with spec files that
already have a correct `Version:` instead of `0`. This makes them more
self-contained and able to build working (s)rpms without the help of
cockpituous.

Support this mode with a new `release-srpm -V` option. By default,
release-srpm would generate a spec file with `Release: 2` as it assumes
that `XXX-1` was already released. `-V` tells it that the `Version:` is
already correct and it should leave the `Release:` at 1.