release: Enable automatic landing in bodhi
In the past two years there has never been a case where we had to recall
a Cockpit update, and pushing it to stable is an extra manual step that
one must remember. Let this happen automatically now.

Note that a bad update can still be blocked by a single -1.