release: Add option to release from a non-master branch
This is useful for cutting releases from stable branches, which just
receive some cherry-picks. This is more practical and robust for
downstream packaging than having to carry huge patches against dist/.

It's also very difficult to carefully tweak all file time stamps to not
trigger webpack when patching - this approach avoids this entirely.