learn: Add simple limit to amount of training data
The current DBSCAN model tries to do a full matrix of all
distances in the training data. Lets just limit to the most
recent N data items. Otherwise the size of the model
and runtime becomes unwieldly.

Hopefully we can eventually migrate to a more efficient
clustering algorithm or implementation.