release: Make build and home directory world-writable
This is similar to what commit 5fe95e634884 did for the cockpit/tests
container. This makes it easier to work in OpenShift environments with
unprivileged random UIDs.

Also drop the unused /build/rpmbuild directory. This fixes running
release-running in the /build directory, as that needs to be empty for
`git clone` to actually work.