learn: Fix state sharing in NCD code
The NCD code relies on preparing shared state for multiple
parallelized processes to share. With the newer 'loky' parallel
backend in scikit-learn the old code no longer works.

We started to see problems with 'list index out of range'.

This is a more explicit way to share state across processes.