learn: Drop dependency on cockpit/base container
The learn container does not need any packages from the base container,
so base it off fedora:latest directly. This makes it smaller and easier
to upgrade to newer Fedora releases independently from the other
containers.

Also drop the /home/user setup. It's in vain, as the random container
uid cannot access that directory anyway.