tasks: Drop cockpit checkout and npm install
The `npm install` was meant as pre-caching in the past, but it is moot
now: cockpit-tasks runs `git clean` as pretty much the first thing it
does. The webhook container entirely ignores this and does its own
checkout. So let's save these > 400 MB from the docker image.

Also drop the checkout -- cockpit-tasks will do that, it doesn't take
long, and it's never going to be really useful as-is, as it quickly gets
out of date.

Closes #251