release: Add option to sit on failure
If a release fails, the release container stops and makes it hard to
debug the failure or finish the release manually. For that, introduce a
new `-s` (sit) option that will sleep indefinitely instead.

As with a sink the actual release-runner runs in a subprocess, we can't
just pass a variable to the parent; so let the child process remove a
temporary flag file on success.

Enable this option in the webhook.

Closes #255