webhook: Use tests-scan --pull-data option
GitHub webhooks suffer from a race condition: When getting a PR event,
calling the API about that very PR often delivers outdated results.  But
the webhook event already gets the entire `pull_request` object as part
of its payload, so let's just use that instead of querying it from the
API again.

https://github.com/cockpit-project/cockpit/pull/11556 introduced a new
`--pull-data` option to tests-scan, so pass the data to it.

This also finally obsoletes checking for the no-test label, so revert
commit f7918f6af.

Fixes https://github.com/cockpit-project/cockpit/issues/11554