tasks: Adjust install-service for podman
Check if podman is available, and if so, prefer it over docker.

Skip the per-container network setup for now. There is no "podman
network", but let's first see if we actually still need this isolation
with podman.

This allows us to run our CI on RHEL 8 or Fedora 31 based hosts.