sink: Port to Python 3
The previous code happily mixed bytes and strings; untangle that, and
consistently write bytes to stdout, stderr, and the log. We receive
bytes, and don't want to introduce any coding/binary data assumptions.

Drop python 2 backwards compatibility. We don't need it on our Fedora
infracloud and OpenShift container sinks, and on fedorapeople.org (which
runs CentOS 7) we can still use an older checkout.

Closes #306