release: Set up default user
This avoids running stuff as root when running the container locally or
in other environments (such as GitHub actions). But keep the home
directory world-writable and /etc/passwd group writable so that
it keeps working in OpenShift with random user IDs.

Set up the git user/email configuration. This was formerly done through
secrets, but there's no reason for that.

Configure kerberos to use a plain file ccache, so that the container
stops assuming a working kernel keyring.

Do the user setup in a separate layer, so that it's easier/faster to
iterate on.

Closes #343