release: Move to GitHub registry, add build workflow
We only use the container in GitHub workflows now, so it makes sense to
keep the image in the nearby registry. This also avoids running into
pull rate limits on docker.io.

Add a manually triggered "build-release" workflow that will allow us to
conveniently build and push a new image from a branch.

Closes #353