tasks: Add packages and configuration for interactive development
With these, developers can run a cockpit/tasks container pretty much
anywhere to do Cockpit development. This works well on our e2e machines
or for toolbox. Further packages can always be installed with
`docker exec -it -u root`, but these provide the most important tools.

Fix ~user home directory permissions, so that tmux and ssh work.

Configure kerberos to use a file ccache, as nothing else (reliably)
works in a container.

A comfortable and recommended workflow is to start the container in the
background and (re)attach to it through byobu. Add a `DEV` run label
with a comment to show how this works. Do *not* pass in the production
cockpit secrets, developers ought to use their own token in such a
situation.