sink: Fix user name detection fallback
The one way how the first attempt realistically fails is through
an `ENOENT` OSError. Replace the bare except with this tighter catch.

It does not make sense to call `os.getlogin()` inside of a
`os.environ.get()`, as any failure of the former would break the latter
as well. So check `$LOGNAME` separately and first, as recommended by getlogin(3).

If both fail (as in GitHub runner environment), read the name directly
from the passwd database in the usual way.