tasks: Use static data directory in run-local.sh
This makes it easier to find the files while the local deployment is
running. We'll also create new subdirectories for logs/images at some
point.

Also keep the files around after the script finished, to be able to
inspect logs and caches. Clean it up at the start of the script instead
of the end.