tasks: Allow user to fix /dev/kvm permissions
Set up sudoers rule for the default container user to be able to fix the
/dev/kvm permissions. This is useful for running the container in
Travis, Cirrus CI and similar environments, where /dev/kvm is 0660 by
default and thus inaccessible for the tasks container.

Note that this does not change anything for user podman containers, as
the user isn't allowed to change /dev/kvm permissions.