Clean up certificate SANs
Remove the triplication of subjectAltName: Make
tasks/credentials/openssl.cnf's server_ca_extensions the single source
of truth, drop the unnecessary SAN from the AMQP client certificate, and
use it in the image certificate script as well.

This also avoids using /etc/pki/tls/openssl.cnf, which does not exist in
Debian/Ubuntu (i.e. on GitHub workflows), and we don't want to rely on
the system config anyway.