tasks: Move self-test logic from workflow to run-local.sh
It is much easier to run it locally that way.

Drop run-local.sh's daemon mode. Instead, always run the tasks container
as a daemon, and follow the logs at the end. That way, we can do
additional things in between.

Add a `-p` option instead to scan/run unit tests on a given PR.