run-local.sh: Test image-{upload,download}
Create a GitHub token and corresponding htpasswd; the latter is
statically computed, as we don't want to introduce a `htpasswd` program
dependency, and reimplementing `$apr1` is too cumbersome. For that
reason, delay setting up the real GitHub token until the time when the
test touches the PR.

Test that image uploading and downloading works. Use the new
COCKPIT_CA_PEM [1] to point bots to the locally generated CA.

[1] https://github.com/cockpit-project/bots/pull/1615