tasks: Fix race condition in run-local.sh
Wait for the images server to be fully up before calling image-upload.
In GitHub workflow it sometimes starts too late (perhaps because of
container image downloads happening in the background?)

Also add "cockpituous-images" to the list of valid DNS names covered by
the TSL certificate, so that we don't need to jump through hoops to use
curl directly.