release-source: Don't build dist tarball out of tree
Recent cockpit fails with that (that's not how we test it in CI), and
it's making the build unnecessarily non-standard. Do exactly what we do
in CI instead.

Simplify the code to only care about autogen.sh (cockpit) or plain
Makefile (starter-kit-like projects).

Now it is easy and straightforward to see what this code is actually
doing.