tasks: Sleep a bit in between iterations when idle in interactive mode
Commit 4707e42b67c7 made the assumption that every run-queue iteration
will block for a few seconds. With the most recent container
refresh/pika update this does not seem true any more: run-queue now
takes so little time that the whole "30 iterations" cycle is done before
tests run-local.sh are done.

As a quick hack, sleep for 10 seconds between iterations in interactive
terminal mode. A better solution would involve to make this *really*
interactive by allowing run-local.sh on the host to control the
iteration. But fixing that will take a little longer, unblock everyone
else in the meantime.