tasks: Stop iteration when receiving SIGTERM
This will allow us to request an orderly shutdown while still finishing
the current test run.

Use this in run-local.sh when it is done with its tests, so that we
don't have to wait for the remaining iterations which won't do anything
anyway. This speeds up the integration test by a few minutes.