release: Bump FMF plan refs in release-koji
As of today, Fedora dist-git FMF test plans have to specify a git repo
and fixed ref for where to find the tests to run.
(https://github.com/psss/tmt/issues/585)

We don't want to bump this manually on every release, so teach
release-koji to do that automatically for every old version that appears
in plans/*.fmf.

This can be dropped once there is a way to just run the tests from the
upstream tarball in the dist-git, or run any "pre-discover" code.