tasks: Drop RabbitMQ pid file assumption in run-local.sh
Use the official `rabbitmqctl` command to check if AMQP server
initialized. The pid file is specific to our current container image,
and not really what we want to wait for anyway.

This was originally done because calling list_queues hangs for a long
time if the server is not up yet. Wrap it in `timeout` to avoid that.