ansible: Split secrets installation into roles
Split OpenShift and "plain directory" (on cloud instances) parts into
separate roles, and factor out the local secrets archive generation, so
that both of the former can depend on that.

This by itself does not change sync-secrets.yml, but makes the roles
re-usable for launch playbooks.

But it does fix sync-secrets.yml to now correctly work with --limit and
a single host, as the `local-secrets-archive` now correctly delegates to
localhost.