tasks: Update required secrets
We haven't used a custom image-stores on our tasks bots in a long time,
and don't plan to, so drop that. But do mention that the SSH
configuration usually comes along with a key and host fingerprint.

We have a lot more production secrets, but none of them are necessary
for testing a basic runner.