Add a slowest tests panel to CI dashboard
The panel is a bar gauge which lists the slowest tests overall as
filtering is currently not possible. For filtering per project to work
we need to introduce a new metric which has a label for every project we
have. As label_values(top_failures_pct, project) is not the same as
label_values(top_slowest_tests, project).

The second issue is that a bar gauge can't be sorted correctly this is a
Grafana limitation. https://github.com/grafana/grafana/issues/17245