tasks: Increase parallel jobs
With 5 parallel tests, cockpit tests often require more than an hour to
run, which exceeds our Service Level Objective [1]. Increase it to 8.
This reduces the number of running bots, and thus increase the queue
time when there is a lot of demand -- but while there isn't, we'll get
feedback much faster.

Don't increase /tmp tmpfs size - 14 GiB should still be enough even for
~ 15 VM overlays (for tests which start multiple VMs); they are not that
big, and projects which do produce a lot of data (such as
cockpit-composer) need to set `$TEST_OVERLAY_DIR` anyway.

Reduce the number of instances on the e2e and AWS machines to compensate
for increasing the test jobs.

Test on CentOS CI fail a lot with TEST_JOBS > 6, due to timeouts, fork()
failures, and VM boot timeouts, so only marginally increase parallelism
from 5 to 6 there. This should at least be enough to push us below one
hour again. Keep the memory limits, as the requirements were recently
reduced [2].

[1] https://github.com/cockpit-project/cockpit/wiki/DevelopmentPrinciples#our-testsci-error-budget
[2] https://github.com/cockpit-project/cockpit/pull/17122