ansible: Use Fedora CoreOS for webhook
CoreOS has a newer podman which avoids long hangs during the container
startup.  New CoreOS images are built daily, so the instance does not
require immediate upgrading; and if/when we do, it's both safer and
faster with rpm-ostree.

Look up the AMI ID dynamically, and use the latest Fedora CoreOS build. The
fixed RHEL 8 AMI is really old by now, and it's annoying to update that
manually.

Drop the now obsolete "system-upgrade" role from launch-webhook.yml.
Drop the "required_packages" role as well. CoreOS has podman and git
pre-installed, and "python3" can't be installed through it anyway as
Ansible already requires that. Replace that bit with a raw rule.

As launching a Fedora CoreOS instance is now reasonably complex, and we
want to use it for the other workloads as well, move it into a reusable
tasks file.