tasks: Introduce elastic cloud runner mode
Let the cockpit-tasks main runner keep track of whether `run-queue` ever
succeeded (i.e. picked something from the queue and ran it), and
otherwise exit with 100. Tell the systemd unit that 100 is a successful
exit code, to avoid confusing log noise.

Let all cockpit-tasks@*.service run in a specific (instead of the
automatic anonymous) `idle-poweroff.slice`, so that we can react to all
tasks containers shutting down.

When calling the installation script with `$IDLE_POWEROFF`, configure
the slice to automatically power off the machine once all cockpit-tasks
instances exited cleanly (we don't want this on failures, so that we can
ssh in and examine them). Use the `poweroff-immediate` heavy hammer
there, to avoid potential hangs on shutdown -- there is nothing to
rescue from the instance anyway.

Plumb that through the AWS Ansible role and document it.