tasks: Increase npm retries and timeout
The RH internal NPM mirror is a bit brittle, so increase the fetch retry
from the default 2 to 6. This mirrors the "Retry npm install 3 times"
logic that we used to have in Cockpit, but is a more elegant and generic
approach that applies to all our projects. Also double the default
timeout from 5 to 10 minutes.