tasks: Increase PID limit
With our high parallelism, tests tend to run into a lot of errors like
this:
```
  File "bots/machine/machine_core/ssh_connection.py", line 109, in wait_user_login
    return self.execute("! test -f /run/nologin && cat /proc/sys/kernel/random/boot_id", direct=True)
  File "bots/machine/machine_core/ssh_connection.py", line 332, in execute
    proc = subprocess.Popen(command_line, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
  File "/usr/lib64/python3.10/subprocess.py", line 966, in __init__
    self._execute_child(args, executable, preexec_fn, close_fds,
  File "/usr/lib64/python3.10/subprocess.py", line 1775, in _execute_child
    self.pid = _posixsubprocess.fork_exec(
BlockingIOError: [Errno 11] Resource temporarily unavailable
```

in particular with Firefox tests (as Firefox tends to create a lot more
threads than Chromium). Increase the podman PID limit from the default
4096 to 16384 to avoid that.

Unfortunately we can't do that on CentOS CI, as kubernetes defines this
on the node, not by pod. So for now we'll have to keep banning firefox
on CentOS CI.