tasks: Fix RabbitMQ/webhook startup race
tasks/cockpit-tasks-webhook.yaml starts the webhook and the Rabbit
containers at the same time. In most cases, RabbitMQ was not fully
running yet when the webhook container got to calling `tests-scan`, so
that the startup failed immediately with an AMQP connection failure.

Mitigate this by adding a waiting loop for AMQP being ready.

Move scan_tasks() below distributed_queue() as the former now calls the
latter.

Make this reproducible in tasks/run-local.sh by starting the webhook
container in interactive mode, and hand it the `-t` provided token, if
present. This is also a nice way of testing the queue scanning.