tasks: Create new S3 user in run-local.sh
Let's not run this test as minioadmin, but create a "cockpituous" user.
Change the minioadmin password to a random one, to effectively disable
it. We don't ever want to muck around with that manually anyway.

This demonstrates how to set this up on the CLI, for future automation
of production.

See https://docs.min.io/minio/baremetal/security/minio-identity-management/user-management.html