tasks: Factorize store URLs in run-local.sh
This will make future changes (particularly https) easier, and it
clarifies what kind of store we are talking to.

We unfortunately cannot do this in the `mc` container, as that
mysteriously fails to resolve `localhost.localdomain` (local podman in
Fedora 35 works fine). However, we  have to address the store that way
in the tasks container, as the token lookup logic in bots' s3 library
requires a fully qualified host name.