tasks: Enable https in run-local.sh's S3 store
minio can speak https when giving it a certificate [1]. This
demonstrates what to do in production for replacing our remaining
cockpit/image store on cockpit-11.

Add "localhost.localdomain" to our SubjectAltNames.  Install our CA into
the mc container, so that it will trust the minio container over https.

Use our lib.s3 instead of plain curl to list the contents of the store,
so that it will automatically use the correct certificate

This uncovers a bug in image-upload: It only provides `--cacert` with
our Cockpituous CA when uploading to a cockpit/images store, not to a
self-deployed S3 store. This gets fixed in
<https://github.com/cockpit-project/bots/pull/3417>.

[1] https://docs.min.io/minio/baremetal/security/network-encryption/minio-tls.html