ansible: Add minio S3 image service to cockpit-11
This replaces our previous cockpit/images container. This was previously
installed via calling images/install-service directly, not in Ansible.
Now provide an inventory group and role for it, to ensure that we always
deploy it in the same way.

This uses the tasks/s3-keys/ and tasks/server.{pem,key} secrets.