images: Use private data volume for minio image store
We previously handed the bare image cache directory to minio as data
volume. While this works fine for reading existing files through the S3
bucket, it doesn't for writing: Uploaded files get created as directory,
with some metadata and data possibly even split between multiple files.

Untangle that by using a persistent private volume for minio. With that,
the data volume file layout and also their permissions don't matter.
This causes some duplication between the volume and the cache directory
on the host, but we have tons of unused space on cockpit-11, so this
does not matter.