ansible: Add task runners in PSI OpenStack
We recently got a "front-door-ci" project on the rhos-d and rhos-01
OpenStack clouds. The latter's nested KVM support and performance is
good enough for our tasks runners (on par with e2e machines).

Add a "launch-tasks" playbook to the new ansible/psi/ directory for
creating a single instance. We don't have very big flavors there, so
each instance can run just one tasks bot. Create these instances in the
"openstack_tasks" group, so that we can refer to them as an Ansible
hosts group.

Install Fedora 35 on them for the time being, as we don't currently have
CoreOS images in that cloud.

Add dynamic OpenStack inventory for PSI rhos-01.  This will pick up all
hosts in PSI rhos-01 for the inventory.  Extend the tasks maintenance
playbooks to also run in the "openstack_tasks" group. See

    ansible-inventory -i inventory/openstack.yml -v --yaml --list

for available hosts and properties.