workflows: Run tasks integration test as user
Trying to build a Fedora container as root has started to fail in a
"funny" way: dnf immediately crashes with "sd-bus call: Connection reset
by peer".

Building and running the container as user has worked fine for a while,
and we do this everywhere else. It was originally introduced for a
missing slirp4netns package, but that got fixed in Ubuntu long ago.