ansible: Move PSI instances to Fedora CoreOS
This image does not exist by default, and most Fedora images are out of
date. This isn't planned to change anytime soon, so instead import our
own CoreOS image. (Internal service-now ref: RITM1234284) Add a shell
script to do that.

With that we are once again running the same OS on all our clouds.