ansible: Switch AWS CoreOS instances to current stable stream
Our AWS image query hits images from the testing stream as well, where
`rpm-ostree upgrade` is broken:

    error: While pulling fedora/x86_64/coreos/testing-devel: No such branch 'fedora/x86_64/coreos/testing-devel' in repository summary

This also disables zincati (the automatic updates service).

If we are on a non-stable stream after installation, rebase to the
current stable stream, so that we always are up to date after running
the playbook, and get proper automatic updates for the instance's
lifetime.