ansible: Set up image cache in PSI
Currently, downloading images from Linode to PSI is really slow, and it
has to happen on each instance separately. To mitigate that, install a
local S3 image store on the first instance.

We don't have any stable IPs or DNS names there, so we have to manage
that as a dynamic "extra" image store entry. Write that to the cockpit
cache directory and bind-mount it into the tasks container as
~/.config/cockpit-dev/image-stores.