images: Add option to install minio S3 store without TLS
This is useful for local S3 stores on instances with ranom IP addresses
and no DNS name, as these can't use TLS: SubjectAltNames just plainly
disallows IP patterns. It's also not necessary if the store is just used
as a local cache, as opposed to "across the internet".

Provide a `disable_tls` variable to teh corresponding Ansible role.