Drop container script labels
We don't actually use this anywhere, nor can/should developers on their
own machines. These were specific to our e2e setup, dangerous to run if
you don't know what you are doing (as they modify the host), and not
really useful for anything. The READMEs document how to properly deploy
these, and for one-off runs there is both the "how to run in toolbox"
documentation as well as our `make tasks-shell` target.