Replace remaining sink container with simple nginx
The only remaining purpose of what was left over of
sink/sink-centosci.yaml was to view the CI weather report, i.e. expose
tests.html and test-results.db through a HTTP server.

We don't need a custom image for that any more, we can just use the
off-the-shelf nginx container for that. We also don't need any secrets
any more, nor write access to the cockpit-images volume. This gets rid
of our `htpasswd` tasks secret.

Note that this does not provide an automatic directory index. That could
be done with a ConfigMap into /etc/nginx/conf.d/, but we don't really
need that for this single purpose.

Keep the https://logs-https-frontdoor.apps.ocp.ci.centos.org/tests.html
route name to not break people's bookmarks, but rename the container
from "sink" to "ci-weather".