ansible: Move PSI auth to application credential
PSI moved from Kerberos password to PIN+TAN a few months ago. It is
impractical to put that directly into secure.yaml, as it's only valid
for a few minutes. Also, exposing your Kerberos password that way was
evil.

Move to application credentials, which are more manageable: they are
limited to one project and role, with a defined expiry, and easy to
retract.

https://access.redhat.com/documentation/en-us/red_hat_openstack_platform/17.0/html/users_and_identity_management_guide/assembly_application-credentials