tasks: Drop unused packages
git-lfs was introduced in commit 9188f0e0931 as an experiment, but we
didn't use it after all.

origin-clients has been broken in Fedora for a long time
(https://bugzilla.redhat.com/show_bug.cgi?id=2094893), and we don't use
it for anything.

wget is not used by any cockpit related project. It was introduced in
commit ede0a6526 for bootstrapping "Atomic" images, but they got
abandoned years ago.